// src/app/@core/dtos/audit.dto.ts
export interface AuditDto {
    description: string;
    action: string;
    actionCategory: string;
    assetType: string;
    status: string;
    userId: string;
    client: string;
    clientId: string;
    assetId: string;
    assetName: string;
    role: string;
    id: string;
    created: string;
    startDate: string;
    endDate: string;
    time: string;
    createdBy: string;
    email: string
}